+++ modules/health-shards-availability/src/main/java/module-info.java
@@ -1,0 +1,4 @@
+module org.elasticsearch.shardhealth {
+    requires org.elasticsearch.server;
+    requires org.apache.lucene.core;
+}
